requirements.txt
numpy
opencv-python
pdfplumber
matplotlib
torch
torchvision
faiss-cpu